rename/#0: 커맨드객체 패키지 변경
