+++ product/src/main/java/takeoff/logistics_service/msa/product/stock/application/dto/request/PostStockRequestDto.java
@@ -1,6 +1,6 @@
 package takeoff.logistics_service.msa.product.stock.application.dto.request;
 
-import takeoff.logistics_service.msa.product.stock.domain.command.CreateStock;
+import takeoff.logistics_service.msa.product.stock.domain.entity.CreateStock;
 import takeoff.logistics_service.msa.product.stock.domain.entity.StockId;
 
 public record PostStockRequestDto(StockIdRequestDto stockId, Integer quantity) {
