+++ product/src/main/java/takeoff/logistics_service/msa/product/stock/application/dto/request/StockIdRequestDto.java
@@ -1,7 +1,7 @@
 package takeoff.logistics_service.msa.product.stock.application.dto.request;
 
 import java.util.UUID;
-import takeoff.logistics_service.msa.product.stock.domain.command.CreateStockId;
+import takeoff.logistics_service.msa.product.stock.domain.entity.CreateStockId;
 
 public record StockIdRequestDto(UUID productId, UUID hubId) {
