--- product/src/main/java/takeoff/logistics_service/msa/product/stock/domain/command/CreateStock.java
@@ -1,6 +1,0 @@
-package takeoff.logistics_service.msa.product.stock.domain.command;
-
-import takeoff.logistics_service.msa.product.stock.domain.entity.StockId;
-
-public record CreateStock(StockId stockId, Integer quantity) {
-}
