+++ product/src/main/java/takeoff/logistics_service/msa/product/stock/domain/entity/CreateStock.java
@@ -1,0 +1,4 @@
+package takeoff.logistics_service.msa.product.stock.domain.entity;
+
+public record CreateStock(StockId stockId, Integer quantity) {
+}
