+++ product/src/test/java/takeoff/logistics_service/msa/product/stock/application/service/StockServiceConcurrencyTest.java
@@ -14,8 +14,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import takeoff.logistics_service.msa.common.exception.BusinessException;
-import takeoff.logistics_service.msa.product.stock.domain.command.CreateStock;
-import takeoff.logistics_service.msa.product.stock.domain.command.CreateStockId;
+import takeoff.logistics_service.msa.product.stock.domain.entity.CreateStock;
+import takeoff.logistics_service.msa.product.stock.domain.entity.CreateStockId;
 import takeoff.logistics_service.msa.product.stock.domain.entity.Stock;
 import takeoff.logistics_service.msa.product.stock.domain.entity.StockId;
 import takeoff.logistics_service.msa.product.stock.domain.repository.StockRepository;
